+++ Sources/Signal+Operators.swift
@@ -1,4 +1,3 @@
-import Foundation
 import Result
 
 extension Signal {
@@ -29,28 +28,6 @@ extension Signal {
 			return self.observe { event in
 				observer.action(event.mapError(transform))
 			}
-		}
-	}
-
-	/// Maps each value in the signal to a new value, lazily evaluating the
-	/// supplied transformation on the specified scheduler.
-	///
-	/// - important: Unlike `map`, there is not a 1-1 mapping between incoming
-	///              values, and values sent on the returned signal. If
-	///              `scheduler` has not yet scheduled `transform` for
-	///              execution, then each new value will replace the last one as
-	///              the parameter to `transform` once it is finally executed.
-	///
-	/// - parameters:
-	///   - transform: The closure used to obtain the returned value from this
-	///                signal's underlying value.
-	///
-	/// - returns: A signal that sends values obtained using `transform` as this
-	///            signal sends values.
-	public func lazyMap<U>(on scheduler: Scheduler, transform: @escaping (Value) -> U) -> Signal<U, Error> {
-		return flatMap(.latest) { value in
-			return SignalProducer({ transform(value) })
-				.start(on: scheduler)
 		}
 	}
 
@@ -338,61 +315,9 @@ extension Signal {
 			}
 		}
 	}
-
-	/// Forward all events onto the given scheduler, instead of whichever
-	/// scheduler they originally arrived upon.
-	///
-	/// - parameters:
-	///   - scheduler: A scheduler to deliver events on.
-	///
-	/// - returns: A signal that will yield `self` values on provided scheduler.
-	public func observe(on scheduler: Scheduler) -> Signal<Value, Error> {
-		return Signal { observer in
-			return self.observe { event in
-				scheduler.schedule {
-					observer.action(event)
-				}
-			}
-		}
-	}
 }
 
 extension Signal {
-	/// Delay `value` and `completed` events by the given interval, forwarding
-	/// them on the given scheduler.
-	///
-	/// - note: failed and `interrupted` events are always scheduled
-	///         immediately.
-	///
-	/// - precondition: `interval` must be non-negative number.
-	///
-	/// - parameters:
-	///   - interval: Interval to delay `value` and `completed` events by.
-	///   - scheduler: A scheduler to deliver delayed events on.
-	///
-	/// - returns: A signal that will delay `value` and `completed` events and
-	///            will yield them on given scheduler.
-	public func delay(_ interval: TimeInterval, on scheduler: DateScheduler) -> Signal<Value, Error> {
-		precondition(interval >= 0)
-
-		return Signal { observer in
-			return self.observe { event in
-				switch event {
-				case .failed, .interrupted:
-					scheduler.schedule {
-						observer.action(event)
-					}
-
-				case .value, .completed:
-					let date = scheduler.currentDate.addingTimeInterval(interval)
-					scheduler.schedule(after: date) {
-						observer.action(event)
-					}
-				}
-			}
-		}
-	}
-
 	/// Skip first `count` number of values then act as usual.
 	///
 	/// - precondition: `count` must be non-negative number.
@@ -902,239 +827,7 @@ extension Signal {
 			}
 		}
 	}
-}
 
-extension Signal {
-	/// Forward the latest value on `scheduler` after at least `interval`
-	/// seconds have passed since *the returned signal* last sent a value.
-	///
-	/// If `self` always sends values more frequently than `interval` seconds,
-	/// then the returned signal will send a value every `interval` seconds.
-	///
-	/// To measure from when `self` last sent a value, see `debounce`.
-	///
-	/// - seealso: `debounce`
-	///
-	/// - note: If multiple values are received before the interval has elapsed,
-	///         the latest value is the one that will be passed on.
-	///
-	/// - note: If `self` terminates while a value is being throttled, that
-	///         value will be discarded and the returned signal will terminate
-	///         immediately.
-	///
-	/// - note: If the device time changed backwards before previous date while
-	///         a value is being throttled, and if there is a new value sent,
-	///         the new value will be passed anyway.
-	///
-	/// - precondition: `interval` must be non-negative number.
-	///
-	/// - parameters:
-	///   - interval: Number of seconds to wait between sent values.
-	///   - scheduler: A scheduler to deliver events on.
-	///
-	/// - returns: A signal that sends values at least `interval` seconds
-	///            appart on a given scheduler.
-	public func throttle(_ interval: TimeInterval, on scheduler: DateScheduler) -> Signal<Value, Error> {
-		precondition(interval >= 0)
-
-		return Signal { observer in
-			let state: Atomic<ThrottleState<Value>> = Atomic(ThrottleState())
-			let schedulerDisposable = SerialDisposable()
-
-			let disposable = CompositeDisposable()
-			disposable += schedulerDisposable
-
-			disposable += self.observe { event in
-				guard let value = event.value else {
-					schedulerDisposable.inner = scheduler.schedule {
-						observer.action(event)
-					}
-					return
-				}
-
-				var scheduleDate: Date!
-				state.modify {
-					$0.pendingValue = value
-
-					let proposedScheduleDate: Date
-					if let previousDate = $0.previousDate, previousDate.compare(scheduler.currentDate) != .orderedDescending {
-						proposedScheduleDate = previousDate.addingTimeInterval(interval)
-					} else {
-						proposedScheduleDate = scheduler.currentDate
-					}
-
-					switch proposedScheduleDate.compare(scheduler.currentDate) {
-					case .orderedAscending:
-						scheduleDate = scheduler.currentDate
-
-					case .orderedSame: fallthrough
-					case .orderedDescending:
-						scheduleDate = proposedScheduleDate
-					}
-				}
-
-				schedulerDisposable.inner = scheduler.schedule(after: scheduleDate) {
-					let pendingValue: Value? = state.modify { state in
-						defer {
-							if state.pendingValue != nil {
-								state.pendingValue = nil
-								state.previousDate = scheduleDate
-							}
-						}
-						return state.pendingValue
-					}
-
-					if let pendingValue = pendingValue {
-						observer.send(value: pendingValue)
-					}
-				}
-			}
-
-			return disposable
-		}
-	}
-
-	/// Conditionally throttles values sent on the receiver whenever
-	/// `shouldThrottle` is true, forwarding values on the given scheduler.
-	///
-	/// - note: While `shouldThrottle` remains false, values are forwarded on the
-	///         given scheduler. If multiple values are received while
-	///         `shouldThrottle` is true, the latest value is the one that will
-	///         be passed on.
-	///
-	/// - note: If the input signal terminates while a value is being throttled,
-	///         that value will be discarded and the returned signal will
-	///         terminate immediately.
-	///
-	/// - note: If `shouldThrottle` completes before the receiver, and its last
-	///         value is `true`, the returned signal will remain in the throttled
-	///         state, emitting no further values until it terminates.
-	///
-	/// - parameters:
-	///   - shouldThrottle: A boolean property that controls whether values
-	///                     should be throttled.
-	///   - scheduler: A scheduler to deliver events on.
-	///
-	/// - returns: A signal that sends values only while `shouldThrottle` is false.
-	public func throttle<P: PropertyProtocol>(while shouldThrottle: P, on scheduler: Scheduler) -> Signal<Value, Error>
-		where P.Value == Bool
-	{
-		return Signal { observer in
-			let initial: ThrottleWhileState<Value> = .resumed
-			let state = Atomic(initial)
-			let schedulerDisposable = SerialDisposable()
-
-			let disposable = CompositeDisposable()
-			disposable += schedulerDisposable
-
-			disposable += shouldThrottle.producer
-				.skipRepeats()
-				.startWithValues { shouldThrottle in
-					let valueToSend = state.modify { state -> Value? in
-						guard !state.isTerminated else { return nil }
-
-						if shouldThrottle {
-							state = .throttled(nil)
-						} else {
-							defer { state = .resumed }
-
-							if case let .throttled(value?) = state {
-								return value
-							}
-						}
-
-						return nil
-					}
-
-					if let value = valueToSend {
-						schedulerDisposable.inner = scheduler.schedule {
-							observer.send(value: value)
-						}
-					}
-			}
-
-			disposable += self.observe { event in
-				let eventToSend = state.modify { state -> Event? in
-					switch event {
-					case let .value(value):
-						switch state {
-						case .throttled:
-							state = .throttled(value)
-							return nil
-						case .resumed:
-							return event
-						case .terminated:
-							return nil
-						}
-
-					case .completed, .interrupted, .failed:
-						state = .terminated
-						return event
-					}
-				}
-
-				if let event = eventToSend {
-					schedulerDisposable.inner = scheduler.schedule {
-						observer.action(event)
-					}
-				}
-			}
-
-			return disposable
-		}
-	}
-
-	/// Forward the latest value on `scheduler` after at least `interval`
-	/// seconds have passed since `self` last sent a value.
-	///
-	/// If `self` always sends values more frequently than `interval` seconds,
-	/// then the returned signal will never send any values.
-	///
-	/// To measure from when the *returned signal* last sent a value, see
-	/// `throttle`.
-	///
-	/// - seealso: `throttle`
-	///
-	/// - note: If multiple values are received before the interval has elapsed,
-	///         the latest value is the one that will be passed on.
-	///
-	/// - note: If the input signal terminates while a value is being debounced,
-	///         that value will be discarded and the returned signal will
-	///         terminate immediately.
-	///
-	/// - precondition: `interval` must be non-negative number.
-	///
-	/// - parameters:
-	///   - interval: A number of seconds to wait before sending a value.
-	///   - scheduler: A scheduler to send values on.
-	///
-	/// - returns: A signal that sends values that are sent from `self` at least
-	///            `interval` seconds apart.
-	public func debounce(_ interval: TimeInterval, on scheduler: DateScheduler) -> Signal<Value, Error> {
-		precondition(interval >= 0)
-
-		let d = SerialDisposable()
-
-		return Signal { observer in
-			return self.observe { event in
-				switch event {
-				case let .value(value):
-					let date = scheduler.currentDate.addingTimeInterval(interval)
-					d.inner = scheduler.schedule(after: date) {
-						observer.send(value: value)
-					}
-
-				case .completed, .failed, .interrupted:
-					d.inner = scheduler.schedule {
-						observer.action(event)
-					}
-				}
-			}
-		}
-	}
-}
-
-extension Signal {
 	/// Forward only those values from `self` that have unique identities across
 	/// the set of all values that have been seen.
 	///
@@ -1182,62 +875,6 @@ extension Signal where Value: Hashable {
 	}
 }
 
-private struct ThrottleState<Value> {
-	var previousDate: Date? = nil
-	var pendingValue: Value? = nil
-}
-
-private enum ThrottleWhileState<Value> {
-	case resumed
-	case throttled(Value?)
-	case terminated
-
-	var isTerminated: Bool {
-		switch self {
-		case .terminated:
-			return true
-		case .resumed, .throttled:
-			return false
-		}
-	}
-}
-
-extension Signal {
-	/// Forward events from `self` until `interval`. Then if signal isn't
-	/// completed yet, fails with `error` on `scheduler`.
-	///
-	/// - note: If the interval is 0, the timeout will be scheduled immediately.
-	///         The signal must complete synchronously (or on a faster
-	///         scheduler) to avoid the timeout.
-	///
-	/// - precondition: `interval` must be non-negative number.
-	///
-	/// - parameters:
-	///   - error: Error to send with failed event if `self` is not completed
-	///            when `interval` passes.
-	///   - interval: Number of seconds to wait for `self` to complete.
-	///   - scheudler: A scheduler to deliver error on.
-	///
-	/// - returns: A signal that sends events for at most `interval` seconds,
-	///            then, if not `completed` - sends `error` with failed event
-	///            on `scheduler`.
-	public func timeout(after interval: TimeInterval, raising error: Error, on scheduler: DateScheduler) -> Signal<Value, Error> {
-		precondition(interval >= 0)
-
-		return Signal { observer in
-			let disposable = CompositeDisposable()
-			let date = scheduler.currentDate.addingTimeInterval(interval)
-
-			disposable += scheduler.schedule(after: date) {
-				observer.send(error: error)
-			}
-
-			disposable += self.observe(observer)
-			return disposable
-		}
-	}
-}
-
 extension Signal where Error == NoError {
 	/// Promote a signal that does not generate failures into one that can.
 	///
@@ -1265,32 +902,6 @@ extension Signal where Error == NoError {
 				}
 			}
 		}
-	}
-
-	/// Forward events from `self` until `interval`. Then if signal isn't
-	/// completed yet, fails with `error` on `scheduler`.
-	///
-	/// - note: If the interval is 0, the timeout will be scheduled immediately.
-	///         The signal must complete synchronously (or on a faster
-	///         scheduler) to avoid the timeout.
-	///
-	/// - parameters:
-	///   - interval: Number of seconds to wait for `self` to complete.
-	///   - error: Error to send with `failed` event if `self` is not completed
-	///            when `interval` passes.
-	///   - scheudler: A scheduler to deliver error on.
-	///
-	/// - returns: A signal that sends events for at most `interval` seconds,
-	///            then, if not `completed` - sends `error` with `failed` event
-	///            on `scheduler`.
-	public func timeout<NewError: Swift.Error>(
-		after interval: TimeInterval,
-		raising error: NewError,
-		on scheduler: DateScheduler
-		) -> Signal<Value, NewError> {
-		return self
-			.promoteErrors(NewError.self)
-			.timeout(after: interval, raising: error, on: scheduler)
 	}
 }
